requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.2.0
antlr4-python3-runtime==4.8
appdirs==1.4.4
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
audioread==3.0.0
cachetools==5.2.0
certifi @ file:///croot/certifi_1665076670883/work/certifi
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
-e git+https://github.com/openai/CLIP.git@d50d76daa670286dd6cacf3bcd80b5e4823fc8e1#egg=clip
ConfigArgParse==1.5.3
cycler==0.11.0
Cython==0.29.32
decorator==5.1.1
einops==0.3.0
face-alignment==1.3.5
ffmpeg==1.4
fonttools==4.38.0
frozenlist==1.3.1
fsspec==2022.10.0
ftfy==6.1.1
future==0.18.2
google-auth==2.13.0
google-auth-oauthlib==0.4.6
grpcio==1.50.0
idna==3.4
imageio==2.22.2
importlib-metadata==5.0.0
joblib==1.2.0
kiwisolver==1.4.4
librosa==0.9.2
llvmlite==0.39.1
lpips==0.1.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
multidict==6.0.2
mutagen==1.46.0
networkx==2.6.3
numba==0.56.3
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1653915516269/work
oauthlib==3.2.2
omegaconf==2.1.1
opencv-python==4.6.0.66
packaging==21.3
pandas==1.1.5
Pillow==9.2.0
pooch==1.6.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydub==0.25.1
pynormalize==0.1.4
pyparsing==3.0.9
pysptk==0.1.21
python-dateutil==2.8.2
python-speech-features==0.6
pytorch-lightning==1.2.5
pytz==2022.5
PyWavelets==1.3.0
pyworld==0.3.1
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.4.2
Resemblyzer==0.1.1.dev0
rsa==4.9
scenedetect==0.6.0.3
scikit-image==0.19.3
scikit-learn==1.0.2
scikit-video==1.1.11
scipy==1.7.3
seaborn==0.12.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0
soundfile==0.11.0
-e git+https://github.com/CompVis/taming-transformers.git@24268930bf1dce879235a7fddd0b2355b84d7ea6#egg=taming_transformers
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
test-tube==0.7.5
threadpoolctl==3.1.0
tifffile==2021.11.2
torch==1.10.0
torchaudio==0.10.0
torchmetrics==0.10.0
torchvision==0.11.0
tqdm==4.64.1
typing==3.7.4.3
typing_extensions @ file:///tmp/abs_ben9emwtky/croots/recipe/typing_extensions_1659638822008/work
urllib3==1.26.12
wcwidth==0.2.5
webrtcvad==2.0.10
Werkzeug==2.2.2
yarl==1.8.1
zipp==3.9.0
